Extract TwoFactorAuthenticatable doubles to support
